+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `unset_input_symbols` and `unset_output_symbols` methods to remove the symbol tables attached to a mutable fst.
 - Added `emplace_arc` and `emplace_arc_unchecked` as provided methods to the `MutableFst` trait.
 - Added `set_symts_from_fst` to MutableFst trait to copy the SymbolTable from another `Fst`.
+- Added `print_weights` field to `DrawingConfig` to avoid print weights when desired.
 
 ### Changed
 - Make `KDELTA` public outside of the crate